// packages/package.trpc-routes/src/types/mainSettings.ts
export type MainPageSettings = {
	id: string
	coverType: string
	isDisplayDiscount: boolean
	isDisplayBooking: boolean
	isDisplayPopular: boolean
	isDisplayFAQ: boolean
}

export type HeaderSettings = {
	isDisplayLogo: boolean
	isDisplayName: boolean
	isDisplaySearch: boolean
	isDisplayBooking: boolean
	isDisplayProfileDetails: boolean
	backgroundColor: string
}

export type FooterSettings = {
  id: string
  isDisplayLogo: boolean
  isDisplayName: boolean
  isDisplaySocialBlock: boolean
}

export type ProfileSettings = {
	id: string
	avatarShape: string
}